+++ src/mod/pbkdf2.mod/pbkdf2.c
@@ -85,8 +85,8 @@ static char *pbkdf2_hash(const char *pass, const char *digest_name,
            strlen("$rounds=4294967295$i") + B64_NTOP_CALCULATE_SIZE(saltlen) +
            1 + B64_NTOP_CALCULATE_SIZE(digestlen);
   if ((outlen + 1) > sizeof out) {
-    putlog(LOG_MISC, "*", "PBKDF2 error: outlen %i > sizeof out %i.", outlen,
-           sizeof out);
+    putlog(LOG_MISC, "*", "PBKDF2 error: outlen %i > sizeof out %ld.", outlen,
+           (long)sizeof out);
     return NULL;
   }
   out2 = out;
